* origin/master:
  Added tier, sort_order_within_tier, image_url_med, image_app_url, id

+++ submittable/parse_submittable.rb
@@ -13,10 +13,14 @@ class Performer
   attr_accessor :home_town
   attr_accessor :name
   attr_accessor :wave
+  attr_accessor :tier
+  attr_accessor :sort_order_within_tier
   attr_accessor :genres
 
   # URLs to our CDN in S3
   attr_accessor :image_url
+  attr_accessor :image_url_med
+  attr_accessor :image_app_url
   attr_accessor :song_url
 
   # social URLs
@@ -31,6 +35,7 @@ class Performer
 
   # Submittable URLs for things that we will upload into S3 to store in our own CDN
   attr_accessor :orig_image_url # from submittable
+  attr_accessor :orig_image_app_url # from submittable
   attr_accessor :orig_song_url  # from submittable
 
   def initialize
@@ -40,14 +45,20 @@ class Performer
 
   def to_json
     {
+      name: name,
       code: code,
+      id: code, # junger wanted this too, same as code
       bio: bio,
       forts: forts,
       home_town: home_town,
-      name: name,
       genres: genres,
       wave: wave,
+      tier: tier,
+      sort_order_within_tier: sort_order_within_tier,
+
       image_url: image_url,
+      image_url_med: image_url_med,
+      image_app_url: image_app_url,
       song_url: song_url,
 
       music_url: music_url,
@@ -70,6 +81,8 @@ class ParseSubmittable
   SONG_BUCKET_NAME  = 'treefort-songs'.freeze
   IMAGE_BUCKET_URL = "https://s3-us-west-2.amazonaws.com/#{IMAGE_BUCKET_NAME}/".freeze
   SONG_BUCKET_URL  = "https://s3-us-west-2.amazonaws.com/#{SONG_BUCKET_NAME}/".freeze
+  IMAGE_FORM_IDS = [784411,685181]
+  IMAGE_APP_FORM_IDS = [826341, 828739]
 
   def run
     @opts = Trollop.options do
@@ -107,7 +120,7 @@ class ParseSubmittable
     year_filter = year_categories.collect { |cat| "&category_id=#{cat['category_id']}" }.join
 
     result_page = 1
-    page_size = @opts[:debug] ? 20 : 100
+    page_size = @opts[:debug] ? 20 : 200
     performers = []
     ready_to_start = @opts[:performer_to_start_at].nil?
 
@@ -118,12 +131,14 @@ class ParseSubmittable
       page_count = [2, page_count].min if @opts[:debug]
 
       submissions['items'].each do |submission|
+        puts "Processing: #{submission['title']}"
         # filter by label
         if submission['labels'].nil?
           labels = []
         else
           label_items = submission['labels']['items']
           labels = label_items.collect { |label| "#{label['label_text']}" }
+          labels = labels.map(&:downcase) # so we can call "include" in a case-insensitive way
         end
 
         # if not debugging, pick the "real" artists
@@ -167,28 +182,64 @@ class ParseSubmittable
           p.wave = 0
         end
 
-        p.home_town     = submission_form_field(submission, 'City, State')
-        p.bio           = submission_form_field(submission, 'Description')
-        p.genres        = submission_form_field(submission, 'Genre')
+        if labels.include?('tier 1')
+          p.tier = 1
+        elsif labels.include?('tier 2')
+          p.tier = 2
+        elsif labels.include?('tier 3')
+          p.tier = 3
+        elsif labels.include?('tier 4')
+          p.tier = 4
+        else
+          puts "Warning, no tier specified for #{p.name}"
+          p.tier = 5
+        end
+
+        p.home_town              = submission_form_field(submission, 'City, State')
+        p.bio                    = submission_form_field(submission, 'Description')
+        p.genres                 = submission_form_field(submission, 'Genre')
         p.genres = p.genres.split(',') unless p.genres.nil?
-        p.music_url     = submission_form_field(submission, 'Music')
-        p.video_url     = submission_form_field(submission, 'Video 1')
-        p.website_url   = submission_form_field(submission, 'Website')
-        p.facebook_url  = submission_form_field(submission, 'Facebook')
-        p.twitter_url   = submission_form_field(submission, 'Twitter')
-        p.instagram_url = submission_form_field(submission, 'Instagram')
+        p.music_url              = submission_form_field(submission, 'Music')
+        p.video_url              = submission_form_field(submission, 'Video 1')
+        p.website_url            = submission_form_field(submission, 'Website')
+        p.facebook_url           = submission_form_field(submission, 'Facebook')
+        p.twitter_url            = submission_form_field(submission, 'Twitter')
+        p.instagram_url          = submission_form_field(submission, 'Instagram')
+        p.sort_order_within_tier = submission_form_field(submission, 'Venue ID')
+
+        if p.sort_order_within_tier.nil?  # make un-set values last
+          p.sort_order_within_tier = 999999
+        else
+          p.sort_order_within_tier = p.sort_order_within_tier.to_i
+        end
 
         if submission['files']
-          images = submission['files'].select { |file| file['mime_type'].start_with?('image/jpeg') }
-          songs  = submission['files'].select { |file| file['mime_type'].start_with?('audio/mp3') }
+          images = submission['files'].select     { |file| IMAGE_FORM_IDS.include?(file['form_field_id']) }
+          images_app = submission['files'].select { |file| IMAGE_APP_FORM_IDS.include?(file['form_field_id']) }
+
+          songs  = submission['files'].select { |file| file['mime_type'].start_with?('audio/mp3') || file['mime_type'].start_with?('audio/mpeg')}
           puts "Warning: #{images.count} images for #{p.name}. Using first one." if images.count > 1
           puts "Warning: No images for #{p.name}." if images.count == 0
+
+          puts "Warning: #{images_app.count} app images for #{p.name}. Using first one." if images_app.count > 1
+          puts "Warning: No app images for #{p.name}." if images_app.count == 0
+
           puts "Warning: #{songs.count} songs for #{p.name}. Using first one." if songs.count > 1
           puts "Warning: No songs for #{p.name}." if songs.count == 0
 
           if images.count == 1
             p.orig_image_url = HOST + images.first['url']
             p.image_url = "#{IMAGE_BUCKET_URL}#{p.code}.jpg"
+            p.image_url_med = "#{IMAGE_BUCKET_URL}#{p.code}-med.jpg"
+          end
+          if images_app.count == 1
+            p.orig_image_app_url = HOST + images_app.first['url']
+            p.image_app_url = "#{IMAGE_BUCKET_URL}#{p.code}-app.jpg"
+          end
+          if images.count == 1
+            p.orig_image_url = HOST + images.first['url']
+            p.image_url = "#{IMAGE_BUCKET_URL}#{p.code}.jpg"
+            p.image_url_med = "#{IMAGE_BUCKET_URL}#{p.code}-med.jpg"
           end
           if songs.count == 1
             p.orig_song_url = HOST + songs.first['url']
@@ -200,7 +251,7 @@ class ParseSubmittable
         end
 
         performers << p
-        puts "Performer: #{p.name}"
+        puts "Performer: #{p.name} sort #{p.sort_order_within_tier}"
       end
       result_page += 1
       break if result_page > page_count
@@ -228,9 +279,57 @@ class ParseSubmittable
 
     performers.each do |p|
       if p.orig_image_url
-        response = CurbFu.get(p.orig_image_url)
+        bucket_key = "#{p.code}.jpg"
+        begin
+          if image_bucket.object(bucket_key).exists?
+            puts "Skipping existing large image. Already in S3 for performer #{p.name}"
+          else
+            response = CurbFu.get(p.orig_image_url)
+            if response.status == 200
+
+              File.open("/tmp/#{bucket_key}", 'wb') do |f|
+                f.binmode
+                f.write response.body
+                f.close
+              end
+              puts "Writing large image to S3 for performer #{p.name}"
+              image_bucket.object(bucket_key).upload_file("/tmp/#{bucket_key}")
+              p.image_url = "#{IMAGE_BUCKET_URL}#{bucket_key}"
+            else
+              puts "Couldn't download image for performer #{p.name}: #{p.orig_image_url}"
+            end
+          end
+        rescue Aws::S3::Errors::ServiceError => error
+          puts 'Unable to add image'
+          puts "#{error.message}"
+        end
+      end
+
+      # if there is a medium size image locally (created by @gbernhardt batch process right now, upload it
+      # as well
+      begin
+        bucket_key = "#{p.code}-med.jpg"
+        if image_bucket.object(bucket_key).exists?
+          puts "Skipping existing medium image. Already in S3 for performer #{p.name}"
+        else
+          if File.exists?("/tmp/#{bucket_key}")
+            puts "Writing medium image to S3 for performer #{p.name}"
+            image_bucket.object(bucket_key).upload_file("/tmp/#{bucket_key}")
+            p.image_url_med = "#{IMAGE_BUCKET_URL}#{bucket_key}"
+          else
+            puts "Warning, No medium image locally for performer #{p.name}"
+          end
+        end
+      rescue Aws::S3::Errors::ServiceError => error
+        puts 'Unable to add image'
+        puts "#{error.message}"
+      end
+
+      if p.orig_image_app_url
+        response = CurbFu.get(p.orig_image_app_url)
         if response.status == 200
-          bucket_key = "#{p.code}.jpg"
+          bucket_key = "#{p.code}-app.jpg"
+
           File.open("/tmp/#{bucket_key}", 'wb') do |f|
             f.binmode
             f.write response.body
@@ -238,33 +337,45 @@ class ParseSubmittable
           end
 
           begin
-            puts "Writing image to S3 for performer #{p.name}"
-            image_bucket.object(bucket_key).upload_file("/tmp/#{bucket_key}")
-            p.image_url = "#{IMAGE_BUCKET_URL}#{bucket_key}"
+            if image_bucket.object(bucket_key).exists?
+              puts "Skipping existing app image. Already in S3 for performer #{p.name}"
+            else
+              puts "Writing app image to S3 for performer #{p.name}"
+              image_bucket.object(bucket_key).upload_file("/tmp/#{bucket_key}")
+              p.image_app_url = "#{IMAGE_BUCKET_URL}#{bucket_key}"
+            end
           rescue Aws::S3::Errors::ServiceError => error
             puts 'Unable to add image'
             puts "#{error.message}"
           end
+        else
+          puts "Couldn't download app image for performer #{p.name}: #{p.orig_image_app_url}"
         end
       end
 
       if p.orig_song_url
-        response = CurbFu.get(p.orig_song_url)
-        if response.status == 200
-          bucket_key = "#{p.code}.mp3"
-          File.open("/tmp/#{bucket_key}", 'wb') do |f|
-            f.binmode
-            f.write response.body
-            f.close
-          end
+        bucket_key = "#{p.code}.mp3"
+        if song_bucket.object(bucket_key).exists?
+          puts "Skipping existing song. Already in S3 for performer #{p.name}"
+        else
+          response = CurbFu.get(p.orig_song_url)
+          if response.status == 200
+            File.open("/tmp/#{bucket_key}", 'wb') do |f|
+              f.binmode
+              f.write response.body
+              f.close
+            end
 
-          begin
-            puts "Writing song  to S3 for performer #{p.name}"
-            song_bucket.object(bucket_key).upload_file("/tmp/#{bucket_key}")
-            p.song_url = "#{SONG_BUCKET_URL}#{bucket_key}"
-          rescue Aws::S3::Errors::ServiceError => error
-            puts 'Unable to add song'
-            puts "#{error.message}"
+            begin
+              puts "Writing song  to S3 for performer #{p.name}"
+              song_bucket.object(bucket_key).upload_file("/tmp/#{bucket_key}")
+              p.song_url = "#{SONG_BUCKET_URL}#{bucket_key}"
+            rescue Aws::S3::Errors::ServiceError => error
+              puts 'Unable to add song'
+              puts "#{error.message}"
+            end
+          else
+            puts "Couldn't download song for performer #{p.name}: #{p.orig_song_url}"
           end
         end
       end
